Added embedded resource support
